[registry-packages-proxy] Move CLI repo resolution into its own file
- Verbatim move of `editionSegments`, `cliRegistryRepository` and
  `countPathSegments` to cli_repo.go, with their test alongside.
- No behavior change; clears room for the CLI root probing logic.
