+++ go_lib/registry-packages-proxy/proxy/cli_handler_test.go
@@ -576,44 +576,6 @@ func TestCLIHandler_GetManifest_NotFound(t *testing.T) {
 	assert.Equal(t, http.StatusNotFound, resp.StatusCode)
 }
 
-func TestCLIRegistryRepository(t *testing.T) {
-	t.Parallel()
-
-	cases := map[string]string{
-		// Edition repositories collapse to the shared root.
-		"registry.deckhouse.io/deckhouse/ee":            "registry.deckhouse.io/deckhouse",
-		"registry.deckhouse.io/deckhouse/ce":            "registry.deckhouse.io/deckhouse",
-		"registry.deckhouse.io/deckhouse/be":            "registry.deckhouse.io/deckhouse",
-		"registry.deckhouse.io/deckhouse/se":            "registry.deckhouse.io/deckhouse",
-		"registry.deckhouse.io/deckhouse/se-plus":       "registry.deckhouse.io/deckhouse",
-		"registry.deckhouse.io/deckhouse/fe/":           "registry.deckhouse.io/deckhouse",
-		"registry.company.com:5000/mirror/deckhouse/ee": "registry.company.com:5000/mirror/deckhouse",
-		// Repositories without an edition segment are the root already.
-		"dev-registry.deckhouse.io/sys/deckhouse-oss": "dev-registry.deckhouse.io/sys/deckhouse-oss",
-		"registry.company.com/deckhouse":              "registry.company.com/deckhouse",
-		"registry.company.com/ee-mirror":              "registry.company.com/ee-mirror",
-		"registry.company.com/deckhouse/EE":           "registry.company.com/deckhouse/EE",
-		"registry.local:5000":                         "registry.local:5000",
-		// CSE keeps its editionless artifacts (the installer) under
-		// deckhouse/cse, so that path is a root of its own.
-		"registry-cse.deckhouse.ru/deckhouse/cse": "registry-cse.deckhouse.ru/deckhouse/cse",
-		// Stripping never eats the last path segment and leaves a bare host,
-		// nor empties the repository outright.
-		"registry.example.com/ee": "registry.example.com/ee",
-		"registry.local/se":       "registry.local/se",
-		"/ee":                     "/ee",
-		"//ee":                    "//ee",
-	}
-
-	for in, want := range cases {
-		in, want := in, want
-		t.Run(in, func(t *testing.T) {
-			t.Parallel()
-			assert.Equal(t, want, cliRegistryRepository(in))
-		})
-	}
-}
-
 // TestCLIHandler_ResolvesArtifactsAtRegistryRoot: every /v1/images/* route
 // reads from the registry root above the cluster's edition repository, for
 // both deckhouse-cli and its plugins.
